+++ controllers/AdminController.php
@@ -4,8 +4,15 @@ namespace humhub\modules\verified\controllers;
 
 use humhub\modules\admin\components\Controller;
 use humhub\modules\verified\models\ConfigureForm;
+use humhub\modules\verified\models\PaymentOptions;
+use humhub\modules\verified\Module;
 use Yii;
 
+/**
+ * AdminController
+ *
+ * @property Module $module
+ */
 class AdminController extends Controller
 {
 
@@ -21,8 +28,16 @@ class AdminController extends Controller
         return $this->render('index', ['model' => $model]);
     }
 
-    public function actionHelp()
+    public function actionOptions()
     {
-        return $this->render('help', []);
+
+        $model = new PaymentOptions();
+        $model->loadSettings();
+
+        if ($model->load(Yii::$app->request->post()) && $model->saveSettings()) {
+            $this->view->saved();
+        }
+
+        return $this->render('options', ['model' => $model]);
     }
 }